+++ src/consts/consts.lua
@@ -5,4 +5,6 @@ return {
     FONT_TITLE_SIZE = 64,
     FONT_HEADER_SIZE = 28,
     FONT_SUB_SIZE = 14,
+    SHAKE_DURATION = 0.3,
+    SHAKE_MAGNITUDE = 5,
 }